requirements.txt
asyncio
python-dotenv
faststream[kafka]